Adding e2e tests for sw-build and renaming globPatterns to staticFileGlobs
