--- packages/sw-build/test/static/example-project-1/page-2.html
@@ -1,1 +1,0 @@
-<h1>Page 2</h1>
