--- packages/sw-build/test/static/example-project-1/styles/stylesheet-1.css
@@ -1,3 +1,0 @@
-body {
-  background-color: red;
-}
